+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.3.1 (Next)
 - No input target paths now defaults to current working directory instead of error.
 - Additionally check `Vagrantfile` and `gemspec` files.
+- Fix behavior for when optional dependency `octocatalog-diff` is absent.
 
 ### 2.3.0
 - Minimum Ruby version increased to 2.7.